Moved to double time and lamda

+++ src/movable/Circle.java
@@ -34,7 +34,7 @@ public class Circle implements Movable{
 	 * @see          movable.Point#moveDown(double)
 	 */
 	@Override
-	public Circle moveDown(final int t, final double factor){
+	public Circle moveDown(final double t, final double factor){
 		this.point.moveDown(t, factor);
 		return this;
 	}
@@ -44,7 +44,7 @@ public class Circle implements Movable{
 	 * @see          movable.Point#moveLeft(double)
 	 */
 	@Override
-	public Circle moveLeft(final int t, final double factor){
+	public Circle moveLeft(final double t, final double factor){
 		this.point.moveLeft(t, factor);
 		return this;
 	}
@@ -54,7 +54,7 @@ public class Circle implements Movable{
 	 * @see          movable.Point#moveRight(double)
 	 */
 	@Override
-	public Circle moveRight(final int t, final double factor){
+	public Circle moveRight(final double t, final double factor){
 		this.point.moveRight(t, factor);
 		return this;
 	}
@@ -65,7 +65,7 @@ public class Circle implements Movable{
 	 * @see          movable.Point#moveUp(double)
 	 */
 	@Override
-	public Circle moveUp(final int t, final double factor){
+	public Circle moveUp(final double t, final double factor){
 		this.point.moveUp(t, factor);
 		return this;
 	}
